requirements.txt
streamlit
python-barcode==0.15.1
Pillow==9.4.0
